Add internal support for multiple returns #565
This update adds support interally for multiple return values.  This
includes support for function/method types with multiple returns, and a
significant amount of the necessary logic.  However, to ensure it
compiles, the code does in some places assume only one return can ever
be provided.  The next step in this process is to add proper support for
multiple return types.